Ran, but inaccurate!
I think it could still use some more training data. Tested accuracy on an image of a snake. The model said there was no snake when given an image with two snakes. The accuracy score is pretty low, so I think this could be why it incorrectly identified no snake.